pyproject.toml
[project]
name = "pghaicker"
dynamic = ["version"]
description = "pghaicker"
authors = [{ name = "Florents Tselai", email = "florents.tselai@gmail.com" }]
license = "MIT"
dependencies = [
    "google-genai",
    "pypandoc"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black"
]

[tool.setuptools.dynamic]
version = {attr = "pghaicker.__version__"}

[project.scripts]
pghaicker = "pghaicker.cli:cli"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

